Move repeat up to array mixin
